+++ NEWS.md
@@ -1,4 +1,4 @@
-# Changes since last release
+# [kwb.utils 0.14.0](https://github.com/KWB-R/kwb.utils/releases/tag/v0.14.0) <small>2023-04-15</small>
 
 - add assignObjects()
 - add catchWarning()
@@ -13,6 +13,11 @@
 - add stopFormatted()
 - improve implementation of hsAddMissingCols()
 - export noSuchElements()
+- improve extractSubstring() 
+- subexpressionMatches(): rename arg "regularExpression" to "pattern", improve
+- add nameElements()
+- parallelNonNA(): use class of input vector "a" as class of output vector
+- fix bug in catAndRun(): show an appropriate time unit, not always "seconds"
 
 # [kwb.utils 0.13.0](https://github.com/KWB-R/kwb.utils/releases/tag/v0.13.0) <small>2022-06-04</small>